Only delete files that have been read to eof, else checkpoint them.

+++ operator/builtin/input/file/file.go
@@ -143,44 +143,50 @@ func (f *InputOperator) poll(ctx context.Context) {
 	wg.Wait()
 
 	if f.deleteAfterRead {
-		f.Debug("cleaning up log files that have been consumed")
+
+		f.Debug("cleaning up log files that have been fully consumed")
+		unfinishedReaders := make([]*Reader, 0, len(readers))
 		for _, reader := range readers {
 			reader.Close()
-			if err := os.Remove(reader.file.Name()); err != nil {
-				f.Errorf("could not delete %s", reader.file.Name())
+			if reader.eof {
+				if err := os.Remove(reader.file.Name()); err != nil {
+					f.Errorf("could not delete %s", reader.file.Name())
+				}
+			} else {
+				unfinishedReaders = append(unfinishedReaders, reader)
 			}
 		}
+		readers = unfinishedReaders
 
-		// Since files are being deleted, skip further tracking
-		return
-	}
+	} else {
 
-	// Detect files that have been rotated out of matching pattern
-	lostReaders := make([]*Reader, 0, len(f.lastPollReaders))
-OUTER:
-	for _, oldReader := range f.lastPollReaders {
-		for _, reader := range readers {
-			if reader.Fingerprint.StartsWith(oldReader.Fingerprint) {
-				continue OUTER
+		// Detect files that have been rotated out of matching pattern
+		lostReaders := make([]*Reader, 0, len(f.lastPollReaders))
+	OUTER:
+		for _, oldReader := range f.lastPollReaders {
+			for _, reader := range readers {
+				if reader.Fingerprint.StartsWith(oldReader.Fingerprint) {
+					continue OUTER
+				}
 			}
+			lostReaders = append(lostReaders, oldReader)
 		}
-		lostReaders = append(lostReaders, oldReader)
-	}
 
-	for _, reader := range lostReaders {
-		wg.Add(1)
-		go func(r *Reader) {
-			defer wg.Done()
-			r.ReadToEnd(ctx)
-		}(reader)
-	}
-	wg.Wait()
+		for _, reader := range lostReaders {
+			wg.Add(1)
+			go func(r *Reader) {
+				defer wg.Done()
+				r.ReadToEnd(ctx)
+			}(reader)
+		}
+		wg.Wait()
 
-	for _, reader := range f.lastPollReaders {
-		reader.Close()
-	}
+		for _, reader := range f.lastPollReaders {
+			reader.Close()
+		}
 
-	f.lastPollReaders = readers
+		f.lastPollReaders = readers
+	}
 
 	f.saveCurrent(readers)
 	f.syncLastPollFiles()
